pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ajebal-daera-translator"
version = "0.1.0"
description = "VRChat real-time speech translation pipeline (STT → LLM → OSC)"
readme = "README.md"
requires-python = ">=3.12,<3.14"
dependencies = [
  "numpy>=2.0",
  "janus>=2.0",
  "cryptography>=41.0",
  "python-osc>=1.9.0",
  "sounddevice>=0.5.0",
  "onnxruntime>=1.18.0",
  "google-genai>=1.56.0",
  "grpcio",
  "dashscope",
  "websockets",
  "deepgram-sdk>=5.0.0",
  "keyring>=25.0",
  "flet>=0.80,<0.81",
  "flet-desktop>=0.80,<0.81",
  "httpx>=0.27",
]

[project.optional-dependencies]
dev = [
  "pytest>=9.0",
  "pytest-asyncio>=1.3",
  "ruff>=0.6.0",
  "black>=24.4.0",
  "pre-commit>=3.7.0",
]
build = [
  "pyinstaller>=6.0",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = "-q -s"
asyncio_mode = "auto"

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I"]
ignore = ["E402", "E501", "E701"]

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
ajebal_daera_translator = ["data/**/*"]
